refactor: AuthMember 패키지 변경
