+++ backend/src/main/java/harustudy/backend/member/controller/MemberController.java
@@ -1,8 +1,7 @@
 package harustudy.backend.member.controller;
 
-import harustudy.backend.auth.AuthMember;
+import harustudy.backend.auth.dto.AuthMember;
 import harustudy.backend.auth.Authenticated;
-import harustudy.backend.member.domain.Member;
 import harustudy.backend.member.dto.MemberResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
